server/requirements.txt
pyvips
psycopg2-binary
python-dotenv
watchdog
